fix(mysql): decode FLOAT/DOUBLE as IEEE-754 bit patterns (#4427)
* fix(mysql): decode FLOAT/DOUBLE as IEEE-754 bit patterns

MySQL sends FLOAT and DOUBLE on the wire as the raw IEEE-754 bit
pattern. Four decode sites read those bytes with a numeric conversion
(`float32(binary.LittleEndian.Uint32(b))`) instead of a reinterpret, so
the integer value of the bit pattern was stored rather than the number.
A column holding 9.99 recorded as 1.0926057e+09 (FLOAT) and
4.621813488089437e+18 (DOUBLE).

The result-set site is the damaging one: it corrupts at *record* time,
so the wrong value is written into mocks.yaml and survives re-recording.
Replay then asserts against a value the database never returned.

Fixed with math.Float32frombits / math.Float64frombits at each site:

  - wire/.../rowscols/binaryProtocolRowPacket.go  (result-set columns)
  - wire/.../preparedstmt/stmtExecutePacket.go    (FLOAT param)
  - wire/.../queryPacket.go                       (query attributes)

Two consequences of the round trip needed fixing alongside it, because
the decode fix alone does not make a FLOAT column usable:

EncodeBinaryRow type-asserted ce.Value nakedly (`ce.Value.(float32)`).
No serializer preserves Go's numeric types: yaml.v3 resolves an untagged
scalar by shape, so a FLOAT column comes back float64 and an integral
DOUBLE comes back int, and encoding/json gives float64 for everything.
Once FLOAT values decoded correctly, those assertions panicked — and a
panic here tears down the proxy's connection handler, so the application
under replay sees "invalid connection" with nothing pointing at the
cause. Replaced with coerceToInt64/coerceToFloat64, which return a
wrapped error naming the column instead.

paramValueEqual compared a float32 against a float64 by widening. A
FLOAT param is float32 off the wire and float64 out of the mock file,
and float64(float32(9.99)) is 9.989999771118164, so a correctly recorded
FLOAT param never matched itself. Narrowed the two mixed arms to float32
precision; float64/float64 is untouched and keeps full precision.

Tests cover both directions and fail without the fix: bit-pattern decode
for FLOAT and DOUBLE, EncodeBinaryRow against YAML-restored rows
(fractional, whole, negative, zero, and BIGINT UNSIGNED above MaxInt64),
and paramValueEqual across a real YAML round trip.


* fix(mysql): keep big integers intact across the JSON mock round trip

Follow-up to the FLOAT/DOUBLE decode fix on this branch. Making FLOAT
columns decode correctly exposed two more places on the same path where a
value is destroyed by being routed through a type that cannot hold it.

DecodeMocksJSON decoded the MySQL spec with a plain json.Unmarshal.
mysql.Request.Message is an interface{}, so that first pass lands every
packet body in a map[string]any with every number as a float64, and
retypeMySQL* then re-marshals that map into the concrete packet type. A
BIGINT UNSIGNED of 18446744073709551615 came back as
1.8446744073709552e+19 and re-marshalled as 18446744073709552000 — a
literal now larger than MaxUint64, so nothing downstream can recover it.

That is not confined to replay. UpdateMocks is a read-modify-write of the
user's mock file: it decodes through this path, drops unused mocks and
writes the survivors back. One `keploy test --remove-unused-mocks` run
rewrites the wrong number onto disk, permanently, in a file people keep in
git. Decoding with UseNumber keeps the original literal through the
intermediate hop; every JSON mock read funnels through this one function,
so the prune path is covered too.

recoverJSONValue gains a matching uint64 rung, so an integer literal above
MaxInt64 stops degrading to float64, and the JSON and YAML formats now
agree on the Go type (yaml.v3 already yields uint64 here). The rung
decodes the token with encoding/json rather than strconv, staying
symmetric with the rest of the file; behaviour is identical to
ParseUint(s, 10, 64) for every form that reaches it.

coerceToInt64 and coerceValueForFieldType no longer convert a float that
cannot represent an integer exactly. Go leaves an out-of-range float→int
conversion implementation-defined; on amd64 it saturates, so 1e300 and
1.8e19 both became MinInt64 and a corrupt mock replayed as a
plausible-looking wrong number. coerceToInt64 now errors, naming the
column and saying to re-record.

This is a silent-to-loud change: a mock already holding an unrepresentable
value used to replay as the wrong number with the suite green, and now
fails the connection. For a tool whose job is fidelity, an undetectably
wrong row is worse than a named error, and such a value has no original
left to recover.

Also fixed alongside:

  - coerceValueForFieldType had no uint64 arm on the FLOAT/DOUBLE types.
    encoding/json writes a float64 in 'f' format below 1e21, so a DOUBLE
    holding 1e19 is marshalled as an integer literal and recovered as
    uint64 — it then fell through unchanged, contradicting the function's
    own contract and re-serializing the column as an integer.
  - An int64 that does not fit int was narrowed unconditionally, undoing
    the 32-bit rung directly above it. Both that guard and exactInt now
    bound the value *before* converting rather than round-tripping
    through the conversion under test, which is implementation-defined
    when out of range (CodeQL go/incorrect-integer-conversion).
  - The float32 arm of coerceToFloat64 routed through a decimal round
    trip. Unreachable (a float32 only ever arrives for FieldTypeFloat,
    where the encoder narrows back exactly), so it is a plain conversion.
  - Comments describing the encoder's old naked type assertions, which
    this branch removed.

e2e: samples-java#150 adds a numeric_fidelity fixture (FLOAT, DOUBLE,
BIGINT UNSIGNED) behind two parameterised endpoints, so the rows arrive as
a binary-protocol result set rather than a text one. Wired into
mysql_dual_conn with a guard asserting on the recorded artifact, because
record and replay were wrong in the same direction and a report-status
check alone can pass while every recorded row is nonsense.

The fixture needs the fix in BOTH binaries — a record binary without it
writes the bit pattern, a replay binary without it panics on a correctly
recorded FLOAT — so MYSQL_NUMERIC_FIDELITY gates it to the
record_build_replay_build cell, as mysql_auto_port does for port
detection. There is no CLI surface to probe, so the matrix states the
fact rather than a probe guessing at it.

Verified against keploy 1b6ead43 and this branch: the fixture records
1.0926057e+09 / 4.621813488089437e+18 on the former and 9.99 / 9.99 on the
latter, replay goes 4/6 to 6/6, and both storage formats pass.


---------

+++ pkg/agent/proxy/integrations/mysql/replayer/match.go
@@ -1574,7 +1574,20 @@ func paramValueEqual(a, b interface{}, nc *util.NoiseChecker) bool {
 		case float32:
 			return av == bv
 		case float64:
-			return float64(av) == bv
+			// Compare at float32 precision, not float64. One side is
+			// genuinely a float32 (FieldTypeFloat off the wire) and the
+			// other has been through YAML, which writes a float32 in its
+			// shortest 32-bit form ("9.99") and reads it back as
+			// float64(9.99). Widening asks the float32 to carry precision
+			// it never had — float64(float32(9.99)) is 9.989999771118164,
+			// so a correctly recorded FLOAT param never matched itself.
+			//
+			// This is the same direction the int/uint arms below already
+			// take. Narrowing does collide for magnitudes float32 cannot
+			// hold (1e-300 compares equal to 0), but a float32 only ever
+			// enters here from a FieldTypeFloat decode and a MySQL FLOAT
+			// cannot carry those, so no real column reaches the collision.
+			return av == float32(bv)
 		case int:
 			return av == float32(bv)
 		case int32:
@@ -1591,7 +1604,8 @@ func paramValueEqual(a, b interface{}, nc *util.NoiseChecker) bool {
 		case float64:
 			return av == bv
 		case float32:
-			return av == float64(bv)
+			// Narrow, don't widen — see the float32 arm above.
+			return float32(av) == bv
 		case int:
 			return av == float64(bv)
 		case int32:
